pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Pass agent prompt and file and agent will Ralph it"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
[tool.pyright]
extraPaths = ["."]
